requirements.txt
# AirEase Backend Dependencies
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.11  # PostgreSQL driver

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Google Gemini AI
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Date/Time
python-dateutil==2.8.2

# Caching (optional)
# redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
